refactor(js): convert the v2 date pickers to TypeScript
